+++ R/csv.R
@@ -1,10 +1,9 @@
 #' Read CmdStan CSV files into R
 #'
 #' @description `read_cmdstan_csv()` is used internally by CmdStanR to read
-#'   CmdStan's output CSV files into \R. It can
-#'   also be used by CmdStan users as a more flexible and efficient alternative
-#'   to `rstan::read_stan_csv()`. See the **Value** section for details on the
-#'   structure of the returned list.
+#'   CmdStan's output CSV files into \R. It can also be used by CmdStan users as
+#'   a more flexible and efficient alternative to `rstan::read_stan_csv()`. See
+#'   the **Value** section for details on the structure of the returned list.
 #'
 #'   It is also possible to create CmdStanR's fitted model objects directly from
 #'   CmdStan CSV files using the `as_cmdstan_fit()` function.
@@ -22,6 +21,10 @@
 #' @param sampler_diagnostics Works the same way as `variables` but for sampler
 #'   diagnostic variables (e.g., `"treedepth__"`, `"accept_stat__"`, etc.).
 #'   Ignored if the model was not fit using MCMC.
+#' @param format The format for storing the draws or point estimates. The
+#'   default depends on the method used to fit the model. See
+#'   [draws][fit-method-draws] for details, in particular the note about speed
+#'   and memory for models with many parameters.
 #'
 #' @return
 #'
@@ -49,14 +52,16 @@
 #' or their diagonals, depending on the type of metric used.
 #' * `step_size`: A list (one element per chain) of the step sizes used.
 #' * `warmup_draws`:  If `save_warmup` was `TRUE` when fitting the model then a
-#' [`draws_array`][posterior::draws_array] of warmup draws.
-#' * `post_warmup_draws`: A [`draws_array`][posterior::draws_array] of
-#' post-warmup draws.
+#' [`draws_array`][posterior::draws_array] (or different format if `format` is
+#' specified) of warmup draws.
+#' * `post_warmup_draws`: A [`draws_array`][posterior::draws_array] (or
+#' different format if `format` is specified) of post-warmup draws.
 #' * `warmup_sampler_diagnostics`:  If `save_warmup` was `TRUE` when fitting the
-#' model then a [`draws_array`][posterior::draws_array] of warmup draws of the
-#' sampler diagnostic variables.
-#' * `post_warmup_sampler_diagnostics`: A [`draws_array`][posterior::draws_array]
-#' of post-warmup draws of the sampler diagnostic variables.
+#' model then a [`draws_array`][posterior::draws_array] (or different format if
+#' `format` is specified) of warmup draws of the sampler diagnostic variables.
+#' * `post_warmup_sampler_diagnostics`: A
+#' [`draws_array`][posterior::draws_array] (or different format if `format` is
+#' specified) of post-warmup draws of the sampler diagnostic variables.
 #'
 #' For [optimization][model-method-optimize] the returned list also includes the
 #' following components:
@@ -66,8 +71,9 @@
 #' For [variational inference][model-method-variational] the returned list also
 #' includes the following components:
 #'
-#' * `draws`: A [`draws_matrix`][posterior::draws_matrix] of draws from the
-#' approximate posterior distribution.
+#' * `draws`: A [`draws_matrix`][posterior::draws_matrix] (or different format
+#' if `format` is specified) of draws from the approximate posterior
+#' distribution.
 #'
 #' For [standalone generated quantities][model-method-generate-quantities] the
 #' returned list also includes the following components:
@@ -85,7 +91,7 @@
 #' # Creating fitting model objects
 #'
 #' # Create a CmdStanMCMC object from the CSV files
-#' fit2 <- as_cmdstan_mcmc(csv_files)
+#' fit2 <- as_cmdstan_fit(csv_files)
 #' fit2$print("beta")
 #'
 #' # Using read_cmdstan_csv
@@ -117,175 +123,150 @@
 #'
 read_cmdstan_csv <- function(files,
                              variables = NULL,
-                             sampler_diagnostics = NULL) {
+                             sampler_diagnostics = NULL,
+                             format = getOption("cmdstanr_draws_format", NULL)) {
+  valid_draws_formats <- c("draws_array", "array", "draws_matrix", "matrix",
+                           "draws_list", "list", "draws_df", "df", "data.frame")
+  if (!is.null(format) && !(format %in% valid_draws_formats)) {
+    stop("The supplied draws format is not valid.", call. = FALSE)
+  }
   checkmate::assert_file_exists(files, access = "r", extension = "csv")
   metadata <- NULL
   warmup_draws <- list()
-  post_warmup_draws <- list()
-  warmup_sampler_diagnostics_draws <- list()
-  post_warmup_sampler_diagnostics_draws <- list()
-  generated_quantities <- list()
-  variational_draws <- NULL
-  point_estimates <- NULL
+  draws <- list()
+  warmup_sampler_diagnostics <- list()
+  post_warmup_sampler_diagnostics <- list()
   inv_metric <- list()
   step_size <- list()
-  col_types <- NULL
-  col_select <- NULL
-  metadata <- NULL
+  csv_metadata <- list()
   time <- data.frame()
-  not_matching <- c()
+  file_idx <- 0
   for (output_file in files) {
-    if (is.null(metadata)) {
-      metadata <- read_csv_metadata(output_file)
-      if (!is.null(metadata$inv_metric)) {
-        inv_metric[[as.character(metadata$id)]] <- metadata$inv_metric
+    file_idx <- length(csv_metadata) + 1
+    csv_metadata[[file_idx]] <- read_csv_metadata(output_file)
+  }
+  if (file_idx > 1) {
+    check_csv_metadata_matches(csv_metadata)
+  }
+  id <- csv_metadata[[1]]$id
+  if (!is.null(csv_metadata[[1]]$inv_metric)) {
+    inv_metric[[as.character(id)]] <- csv_metadata[[1]]$inv_metric
+  }
+  if (!is.null(csv_metadata[[1]]$step_size_adaptation)) {
+    step_size[[as.character(id)]] <- csv_metadata[[1]]$step_size_adaptation
+  }
+  if (!is.null(csv_metadata[[1]]$time)) {
+    time <- rbind(time, csv_metadata[[1]]$time)
+  }
+  if (length(csv_metadata) > 1) {
+    for (file_id in 2:length(csv_metadata)) {
+      file_metadata <- csv_metadata[[file_id]]
+      id <- file_metadata$id
+      csv_metadata[[1]]$id <- c(csv_metadata[[1]]$id, id)
+      csv_metadata[[1]]$seed <- c(csv_metadata[[1]]$seed, file_metadata$seed)
+      csv_metadata[[1]]$init <- c(csv_metadata[[1]]$init, file_metadata$init)
+      csv_metadata[[1]]$step_size <- c(csv_metadata[[1]]$step_size, file_metadata$step_size)
+      csv_metadata[[1]]$step_size_adaptation <- c(csv_metadata[[1]]$step_size_adaptation, file_metadata$step_size_adaptation)
+      csv_metadata[[1]]$fitted_params <- c(csv_metadata[[1]]$fitted_params, file_metadata$fitted_params)
+      if (!is.null(file_metadata$inv_metric)) {
+        inv_metric[[as.character(id)]] <- file_metadata$inv_metric
       }
-      if (!is.null(metadata$step_size_adaptation)) {
-        step_size[[as.character(metadata$id)]] <- metadata$step_size_adaptation
+      if (!is.null(file_metadata$step_size_adaptation)) {
+        step_size[[as.character(id)]] <- file_metadata$step_size_adaptation
       }
-      if (!is.null(metadata$time)) {
-        time <- rbind(time, metadata$time)
+      if (!is.null(file_metadata$time)) {
+        time <- rbind(time, file_metadata$time)
       }
+    }
+  }
+  metadata <- csv_metadata[[1]]
+  uniq_seed <- unique(metadata$seed)
+  if (length(uniq_seed) == 1) {
+    metadata$seed <- uniq_seed
+  }
+  if (is.null(variables)) { # variables = NULL returns all
+    variables <- metadata$model_params
+  } else if (!any(nzchar(variables))) { # if variables = "" returns none
+    variables <- NULL
+  } else { # filter using variables
+    res <- matching_variables(variables, repair_variable_names(metadata$model_params))
+    if (length(res$not_found)) {
+      stop("Can't find the following variable(s) in the output: ",
+            paste(res$not_found, collapse = ", "), call. = FALSE)
+    }
+    variables <- unrepair_variable_names(res$matching)
+  }
+  if (is.null(sampler_diagnostics)) {
+    sampler_diagnostics <- metadata$sampler_diagnostics
+  } else if (!any(nzchar(sampler_diagnostics))) { # if sampler_diagnostics = "" returns none
+    sampler_diagnostics <- NULL
+  } else {
+    selected_sampler_diag <- rep(FALSE, length(metadata$sampler_diagnostics))
+    not_found <- NULL
+    for (p in sampler_diagnostics) {
+      matches <- metadata$sampler_diagnostics == p | startsWith(metadata$sampler_diagnostics, paste0(p,"."))
+      if (!any(matches)) {
+        not_found <- c(not_found, p)
+      }
+      selected_sampler_diag <- selected_sampler_diag | matches
+    }
+    if (length(not_found)) {
+      stop("Can't find the following sampler diagnostic(s) in the output: ",
+            paste(not_found, collapse = ", "), call. = FALSE)
+    }
+    sampler_diagnostics <- metadata$sampler_diagnostics[selected_sampler_diag]
+  }
+  num_warmup_draws <- ceiling(metadata$iter_warmup / metadata$thin)
+  num_post_warmup_draws <- ceiling(metadata$iter_sampling / metadata$thin)
+  for (output_file in files) {
+    if (os_is_windows()) {
+      grep_path <- repair_path(Sys.which("grep.exe"))
+      fread_cmd <- paste0(grep_path, " -v '^#' --color=never '", output_file, "'")
     } else {
-      csv_file_info <- read_csv_metadata(output_file)
-      check <- check_csv_metadata_matches(metadata, csv_file_info)
-      if (!is.null(check$error)) {
-        stop(check$error, call. = FALSE)
-      }
-      not_matching <- c(not_matching, check$not_matching)
-      metadata$id <- c(metadata$id, csv_file_info$id)
-      metadata$seed <- c(metadata$seed, csv_file_info$seed)
-      metadata$init <- c(metadata$init, csv_file_info$init)
-      metadata$step_size <- c(metadata$step_size, csv_file_info$step_size)
-      metadata$step_size_adaptation <- c(metadata$step_size_adaptation, csv_file_info$step_size_adaptation)
-      metadata$fitted_params <- c(metadata$fitted_params, csv_file_info$fitted_params)
-
-      if (!is.null(csv_file_info$inv_metric)) {
-        inv_metric[[as.character(csv_file_info$id)]] <- csv_file_info$inv_metric
-      }
-      if (!is.null(csv_file_info$step_size_adaptation)) {
-        step_size[[as.character(csv_file_info$id)]] <- csv_file_info$step_size_adaptation
-      }
-      if (!is.null(csv_file_info$time)) {
-        time <- rbind(time, csv_file_info$time)
-      }
+      fread_cmd <- paste0("grep -v '^#' --color=never '", output_file, "'")
     }
-    if (is.null(col_select)) {
-      if (is.null(variables)) { # variables = NULL returns all
-        variables <- metadata$model_params
-      } else if (!any(nzchar(variables))) { # if variables = "" returns none
-        variables <- NULL
-      } else { # filter using variables
-        res <- matching_variables(variables, repair_variable_names(metadata$model_params))
-        if (length(res$not_found)) {
-          stop("Can't find the following variable(s) in the output: ",
-               paste(res$not_found, collapse = ", "), call. = FALSE)
-        }
-        variables <- unrepair_variable_names(res$matching)
-      }
-      if (is.null(sampler_diagnostics)) {
-        sampler_diagnostics <- metadata$sampler_diagnostics
-      } else if (!any(nzchar(sampler_diagnostics))) { # if sampler_diagnostics = "" returns none
-        sampler_diagnostics <- NULL
-      } else {
-        selected_sampler_diag <- rep(FALSE, length(metadata$sampler_diagnostics))
-        not_found <- NULL
-        for (p in sampler_diagnostics) {
-          matches <- metadata$sampler_diagnostics == p | startsWith(metadata$sampler_diagnostics, paste0(p,"."))
-          if (!any(matches)) {
-            not_found <- c(not_found, p)
-          }
-          selected_sampler_diag <- selected_sampler_diag | matches
-        }
-        if (length(not_found)) {
-          stop("Can't find the following sampler diagnostic(s) in the output: ",
-               paste(not_found, collapse = ", "), call. = FALSE)
-        }
-        sampler_diagnostics <- metadata$sampler_diagnostics[selected_sampler_diag]
-      }
-      if (metadata$method == "generate_quantities") {
-        col_select <- c(col_select, variables)
-      } else {
-        col_select <- "lp__"
-        col_select <- c(col_select, variables[variables!="lp__"])
-        col_select <- c(col_select, sampler_diagnostics)
-      }
-    }
-    if (metadata$method == "sample") {
-      num_warmup_draws <- ceiling(metadata$iter_warmup / metadata$thin)
-      num_post_warmup_draws <- ceiling(metadata$iter_sampling / metadata$thin)
-      all_draws <- num_warmup_draws + num_post_warmup_draws
-    } else if (metadata$method == "variational") {
-      all_draws <- metadata$output_samples
-    } else if (metadata$method == "optimize") {
-      all_draws <- 1
-    }
-    if (length(col_select) > 0) {
-      if (os_is_windows()) {
-        grep_path <- repair_path(Sys.which("grep.exe"))
-        fread_cmd <- paste0(grep_path, " -v '^#' --color=never ", output_file)
-      } else {
-        fread_cmd <- paste0("grep -v '^#' --color=never ", output_file)
-      }
+    if (length(sampler_diagnostics) > 0) {
+      post_warmup_sd_id <- length(post_warmup_sampler_diagnostics) + 1
+      warmup_sd_id <- length(warmup_sampler_diagnostics) + 1
       suppressWarnings(
-      draws <- data.table::fread(
+        post_warmup_sampler_diagnostics[[post_warmup_sd_id]] <- data.table::fread(
           cmd = fread_cmd,
-          select = col_select,
+          select = sampler_diagnostics,
           data.table = FALSE
         )
       )
-    } else {
-      draws <- NULL
-    }
-    if (nrow(draws) > 0) {
-      if (metadata$method == "sample") {
-        if (metadata$save_warmup == 1) {
-          if (length(variables) > 0) {
-            warmup_draws[[length(warmup_draws) + 1]] <- draws[1:num_warmup_draws, variables, drop = FALSE]
-            if (num_post_warmup_draws > 0) {
-              post_warmup_draws[[length(post_warmup_draws) + 1]] <- draws[(num_warmup_draws+1):all_draws, variables, drop = FALSE]
-            }
-          }
-          if (length(sampler_diagnostics) > 0) {
-            warmup_sampler_diagnostics_draws[[length(warmup_sampler_diagnostics_draws) + 1]] <- draws[1:num_warmup_draws, sampler_diagnostics, drop = FALSE]
-            if (num_post_warmup_draws > 0) {
-              post_warmup_sampler_diagnostics_draws[[length(post_warmup_sampler_diagnostics_draws) + 1]] <- draws[(num_warmup_draws+1):all_draws, sampler_diagnostics, drop = FALSE]
-            }
-          }
+      if (metadata$method == "sample" && metadata$save_warmup == 1 && num_warmup_draws > 0) {
+        warmup_sampler_diagnostics[[warmup_sd_id]] <-
+          post_warmup_sampler_diagnostics[[post_warmup_sd_id]][1:num_warmup_draws,,drop = FALSE]
+        if (num_post_warmup_draws > 0) {
+          post_warmup_sampler_diagnostics[[post_warmup_sd_id]] <-
+            post_warmup_sampler_diagnostics[[post_warmup_sd_id]][(num_warmup_draws+1):(num_warmup_draws + num_post_warmup_draws),,drop = FALSE]
         } else {
-            warmup_draws <- NULL
-            warmup_sampler_diagnostics_draws <- NULL
-            if (length(variables) > 0) {
-              post_warmup_draws[[length(post_warmup_draws) + 1]] <- draws[, variables, drop = FALSE]
-            }
-            if (length(sampler_diagnostics) > 0 && all(metadata$algorithm != "fixed_param")) {
-              post_warmup_sampler_diagnostics_draws[[length(post_warmup_sampler_diagnostics_draws) + 1]] <- draws[, sampler_diagnostics, drop = FALSE]
-            }
+          post_warmup_sampler_diagnostics[[post_warmup_sd_id]] <- NULL
         }
-      } else if (metadata$method == "variational") {
-        # ignore first line as it's just the mean and lp__ as it's always 0
-        variational_draws <- posterior::as_draws_matrix(
-          draws[-1, colnames(draws) != "lp__", drop=FALSE]
+      }
+    }
+    if (length(variables) > 0) {
+      draws_list_id <- length(draws) + 1
+      warmup_draws_list_id <- length(warmup_draws) + 1
+      suppressWarnings(
+        draws[[draws_list_id]] <- data.table::fread(
+          cmd = fread_cmd,
+          select = variables,
+          data.table = FALSE
         )
-        if ("log_p__" %in% posterior::variables(variational_draws)) {
-          variational_draws <- posterior::rename_variables(variational_draws, lp__ = "log_p__")
+      )
+      if (metadata$method == "sample" && metadata$save_warmup == 1 && num_warmup_draws > 0) {
+        warmup_draws[[warmup_draws_list_id]] <-
+          draws[[draws_list_id]][1:num_warmup_draws,,drop = FALSE]
+        if (num_post_warmup_draws > 0) {
+          draws[[draws_list_id]] <- draws[[draws_list_id]][(num_warmup_draws+1):(num_warmup_draws + num_post_warmup_draws),,drop = FALSE]
+        } else {
+          draws[[draws_list_id]] <- NULL
         }
-        if ("log_g__" %in% posterior::variables(variational_draws)) {
-          variational_draws <- posterior::rename_variables(variational_draws, lp_approx__ = "log_g__")
-        }
-      } else if (metadata$method == "optimize") {
-        point_estimates <- posterior::as_draws_matrix(draws[1,, drop=FALSE])[, variables]
-      } else if (metadata$method == "generate_quantities") {
-        generated_quantities[[length(generated_quantities) + 1]] <- draws
       }
     }
   }
-  if (length(not_matching) > 0) {
-    not_matching_list <- paste(unique(not_matching), collapse = ", ")
-    warning("Supplied CSV files do not match in the following arguments: ",
-            paste(not_matching_list, collapse = ", "), call. = FALSE)
-  }
-
   metadata$inv_metric <- NULL
   metadata$model_params <- repair_variable_names(metadata$model_params)
   repaired_variables <- repair_variable_names(variables)
@@ -297,34 +278,72 @@ read_cmdstan_csv <- function(files,
     repaired_variables <- gsub("log_p__", "lp__", repaired_variables)
     repaired_variables <- gsub("log_g__", "lp_approx__", repaired_variables)
   }
-
   model_param_dims <- variable_dims(metadata$model_params)
   metadata$stan_variable_dims <- model_param_dims
   metadata$stan_variables <- names(model_param_dims)
 
   if (metadata$method == "sample") {
-    warmup_draws <- bind_list_of_draws_array(warmup_draws)
-    if (!is.null(warmup_draws)) {
+    if (is.null(format)) {
+      format <- "draws_array"
+    }
+    as_draws_format <- as_draws_format_fun(format)
+    if (length(warmup_draws) > 0) {
+      warmup_draws <- do.call(as_draws_format, list(warmup_draws))
       posterior::variables(warmup_draws) <- repaired_variables
+      if (posterior::niterations(warmup_draws) == 0) {
+        warmup_draws <- NULL
+      }
+    } else {
+      warmup_draws <- NULL
     }
-    post_warmup_draws <- bind_list_of_draws_array(post_warmup_draws)
-    if (!is.null(post_warmup_draws)) {
-      posterior::variables(post_warmup_draws) <- repaired_variables
+    if (length(draws) > 0) {
+      draws <-  do.call(as_draws_format, list(draws))
+      posterior::variables(draws) <- repaired_variables
+      if (posterior::niterations(draws) == 0) {
+        draws <- NULL
+      }
+    } else {
+      draws <- NULL
     }
-    warmup_sampler_diagnostics_draws <- bind_list_of_draws_array(warmup_sampler_diagnostics_draws)
-    post_warmup_sampler_diagnostics_draws <- bind_list_of_draws_array(post_warmup_sampler_diagnostics_draws)
+    if (length(warmup_sampler_diagnostics) > 0) {
+      warmup_sampler_diagnostics <- do.call(as_draws_format, list(warmup_sampler_diagnostics))
+      if (posterior::niterations(warmup_sampler_diagnostics) == 0) {
+        warmup_sampler_diagnostics <- NULL
+      }
+    } else {
+      warmup_sampler_diagnostics <- NULL
+    }
+    if (length(post_warmup_sampler_diagnostics) > 0) {
+      post_warmup_sampler_diagnostics <- do.call(as_draws_format, list(post_warmup_sampler_diagnostics))
+      if (posterior::niterations(post_warmup_sampler_diagnostics) == 0) {
+        post_warmup_sampler_diagnostics <- NULL
+      }
+    } else {
+      post_warmup_sampler_diagnostics <- NULL
+    }
     list(
       metadata = metadata,
       time = list(total = NA_integer_, chains = time),
       inv_metric = inv_metric,
       step_size = step_size,
       warmup_draws = warmup_draws,
-      post_warmup_draws = post_warmup_draws,
-      warmup_sampler_diagnostics = warmup_sampler_diagnostics_draws,
-      post_warmup_sampler_diagnostics = post_warmup_sampler_diagnostics_draws
+      post_warmup_draws = draws,
+      warmup_sampler_diagnostics = warmup_sampler_diagnostics,
+      post_warmup_sampler_diagnostics = post_warmup_sampler_diagnostics
     )
   } else if (metadata$method == "variational") {
+    if (is.null(format)) {
+      format <- "draws_matrix"
+    }
+    as_draws_format <- as_draws_format_fun(format)
+    variational_draws <- do.call(as_draws_format, list(draws[[1]][-1, colnames(draws[[1]]) != "lp__", drop=FALSE]))
     if (!is.null(variational_draws)) {
+      if ("log_p__" %in% posterior::variables(variational_draws)) {
+        variational_draws <- posterior::rename_variables(variational_draws, lp__ = "log_p__")
+      }
+      if ("log_g__" %in% posterior::variables(variational_draws)) {
+        variational_draws <- posterior::rename_variables(variational_draws, lp_approx__ = "log_g__")
+      }
       posterior::variables(variational_draws) <- repaired_variables
     }
     list(
@@ -332,6 +351,12 @@ read_cmdstan_csv <- function(files,
       draws = variational_draws
     )
   } else if (metadata$method == "optimize") {
+    if (is.null(format)) {
+      format <- "draws_matrix"
+    }
+    as_draws_format <- as_draws_format_fun(format)
+    point_estimates <- do.call(as_draws_format, list(draws[[1]][1,, drop=FALSE]))
+    point_estimates <- posterior::subset_draws(point_estimates, variable = variables)
     if (!is.null(point_estimates)) {
       posterior::variables(point_estimates) <- repaired_variables
     }
@@ -340,13 +365,17 @@ read_cmdstan_csv <- function(files,
       point_estimates = point_estimates
     )
   } else if (metadata$method == "generate_quantities") {
-    generated_quantities <- bind_list_of_draws_array(generated_quantities)
-    if (!is.null(generated_quantities)) {
-      posterior::variables(generated_quantities) <- repaired_variables
+    if (is.null(format)) {
+      format <- "draws_array"
+    }
+    as_draws_format <- as_draws_format_fun(format)
+    draws <- do.call(as_draws_format, list(draws))
+    if (!is.null(draws)) {
+      posterior::variables(draws) <- repaired_variables
     }
     list(
       metadata = metadata,
-      generated_quantities = generated_quantities
+      generated_quantities = draws
     )
   }
 }
@@ -372,8 +401,8 @@ read_sample_csv <- function(files,
 #'   be performed after reading in the files? The default is `TRUE` but set to
 #'   `FALSE` to avoid checking for problems with divergences and treedepth.
 #'
-as_cmdstan_fit <- function(files, check_diagnostics = TRUE) {
-  csv_contents <- read_cmdstan_csv(files)
+as_cmdstan_fit <- function(files, check_diagnostics = TRUE, format = getOption("cmdstanr_draws_format", NULL)) {
+  csv_contents <- read_cmdstan_csv(files, format = format)
   switch(
     csv_contents$metadata$method,
     "sample" = CmdStanMCMC_CSV$new(csv_contents, files, check_diagnostics),
@@ -512,9 +541,9 @@ read_csv_metadata <- function(csv_file) {
   total_time <- 0
   if (os_is_windows()) {
     grep_path <- repair_path(Sys.which("grep.exe"))
-    fread_cmd <- paste0(grep_path, " '^[#a-zA-Z]' --color=never ", csv_file)
+    fread_cmd <- paste0(grep_path, " '^[#a-zA-Z]' --color=never '", csv_file, "'")
   } else {
-    fread_cmd <- paste0("grep '^[#a-zA-Z]' --color=never ", csv_file)
+    fread_cmd <- paste0("grep '^[#a-zA-Z]' --color=never '", csv_file, "'")
   }
   suppressWarnings(
     metadata <- data.table::fread(
@@ -651,55 +680,60 @@ read_csv_metadata <- function(csv_file) {
 #' @noRd
 #' @param a,b Two lists returned by `read_csv_metadata()` to compare.
 #'
-check_csv_metadata_matches <- function(a, b) {
-  if (a$model_name != b$model_name) {
-    return(list(error = "Supplied CSV files were not generated with the same model!"))
+check_csv_metadata_matches <- function(csv_metadata) {
+  model_name <- sapply(csv_metadata, function(x) x$model_name)
+  if (!all(model_name == model_name[1])) {
+    stop("Supplied CSV files were not generated with the same model!", call. = FALSE)
   }
-  if (a$method != b$method) {
-    return(list(error = "Supplied CSV files were produced by different methods and need to be read in separately!"))
+  method <- sapply(csv_metadata, function(x) x$method)
+  if (!all(method == method[1])) {
+    stop("Supplied CSV files were produced by different methods and need to be read in separately!", call. = FALSE)
   }
-  if ((length(a$model_params) != length(b$model_params)) ||
-      !(all(a$model_params == b$model_params) &&
-        all(a$sampler_diagnostics == b$sampler_diagnostics))) {
-    return(list(error = "Supplied CSV files have samples for different variables!"))
-  }
-  if (a$method == "sample") {
-    if (a$iter_sampling != b$iter_sampling ||
-        a$thin != b$thin ||
-        a$save_warmup != b$save_warmup ||
-        (a$save_warmup == 1 && a$iter_warmup != b$iter_warmup)) {
-      return(list(error = "Supplied CSV files do not match in the number of output samples!"))
+  for(i in 2:length(csv_metadata)) {
+    if (length(csv_metadata[[1]]$model_params) != length(csv_metadata[[i]]$model_params) ||
+      !all(csv_metadata[[1]]$model_params == csv_metadata[[i]]$model_params)) {
+      stop("Supplied CSV files have samples for different variables!", call. = FALSE)
     }
-  } else if (a$method == "variational") {
-    if (a$output_samples != b$output_samples) {
-      return(list(error = "Supplied CSV files do not match in the number of output samples!"))
+  }
+  if (method[1] == "sample") {
+    iter_sampling <- sapply(csv_metadata, function(x) x$iter_sampling)
+    thin <- sapply(csv_metadata, function(x) x$thin)
+    save_warmup <- sapply(csv_metadata, function(x) x$save_warmup)
+    iter_warmup <- sapply(csv_metadata, function(x) x$iter_warmup)
+    if (!all(iter_sampling == iter_sampling[1]) ||
+        !all(thin == thin[1]) ||
+        !all(save_warmup == save_warmup[1])||
+        (save_warmup[1] == 1 && !all(iter_warmup == iter_warmup[1]))) {
+      stop("Supplied CSV files do not match in the number of output samples!", call. = FALSE)
+    }
+  } else if (method[1] == "variational") {
+    output_samples <- sapply(csv_metadata, function(x) x$output_samples)
+    if (!all(output_samples == output_samples[1])) {
+      stop("Supplied CSV files do not match in the number of output samples!", call. = FALSE)
     }
   }
   match_list <- c("stan_version_major", "stan_version_minor", "stan_version_patch", "gamma", "kappa",
                   "t0", "init_buffer", "term_buffer", "window", "algorithm", "engine", "max_treedepth",
                   "metric", "stepsize_jitter", "adapt_engaged", "adapt_delta", "iter_warmup")
   not_matching <- c()
-  for (name in names(a)) {
-    if ((name %in% match_list) && (is.null(b[[name]]) ||  all(a[[name]] != b[[name]]))) {
-      not_matching <- c(not_matching, name)
+  for (name in names(csv_metadata[[1]])) {
+    if (name %in% match_list) {
+      values <- sapply(csv_metadata, function(x) x[[name]])
+      if (any(sapply(values, function(x) is.null(x)))) {
+        not_matching <- c(not_matching, name)
+        next
+      }
+      if (!all(values == values[1])) {
+        not_matching <- c(not_matching, name)
+      }
     }
   }
-  list(not_matching = not_matching)
-}
-
-bind_list_of_draws_array <- function(draws, along = "chain") {
-  if (!is.null(draws) && length(draws) > 0) {
-    if (length(draws) > 1) {
-      draws <- lapply(draws, posterior::as_draws_array)
-      draws[["along"]] <- along
-      draws <- do.call(posterior::bind_draws, draws)
-    } else {
-      draws <- posterior::as_draws_array(draws[[1]])
-    }    
-  } else {
-    draws <- NULL
+  if (length(not_matching) > 0) {
+    not_matching_list <- paste(unique(not_matching), collapse = ", ")
+    warning("Supplied CSV files do not match in the following arguments: ",
+            paste(not_matching_list, collapse = ", "), call. = FALSE)
   }
-  draws
+  NULL
 }
 
 # convert names like beta.1.1 to beta[1,1]